+++ helper-scripts/src/stackManagementTool/generateEnv/generateEnvTool.ts
@@ -3,7 +3,7 @@ import {
   whichStacksToGenerateEnvFor,
 } from "./prompts.js";
 import { echo } from "zx";
-import { generateEnvs } from "./getStackName/generateEnvs/generateEnvs.js";
+import { generateEnvs } from "./generateEnvs/generateEnvs.js";
 import { getStackName } from "./getStackName/getStackName.js";
 
 export const generateEnvTool = async (): Promise<void> => {
