+++ helper-scripts/src/stackManagementTool/generateEnv/generateEnvs/prompts.ts
