+++ helper-scripts/src/stackManagementTool/protectedStacks/protectedStacks.ts
@@ -2,6 +2,9 @@ export const protectedStacks = [
   "mob-sts-mock",
   "mob-sts-mock-tir",
   "mob-sts-mock-pl",
+  "mob-test-resources",
+  "mob-test-resources-tir",
+  "mob-test-resources-pl",
   "mob-async-backend",
   "mob-async-backend-tir",
   "mob-async-backend-pl",
